modified:   CString/CString.h

+++ CString/CString.h
@@ -125,7 +125,7 @@ int Cmem<T>::memcmp(const void *s1, const void *s2, unsigned int n)
 	int j;
 	for(i=0;i<n;i++)
 	{
-		j=static_cast<int>((static_cast<ubyte *>(s1))[i]-(static_cast<ubyte *>(s2))[i]);
+		j=static_cast<int>(static_cast<byte>((static_cast<ubyte *>(s1))[i]-(static_cast<ubyte *>(s2))[i]));
 		if(j) return j;
 	}
 return 0;
